Refactor rte_list functions to static methods
